google auth in, logout button works, removed logout page because it was redundant
