+++ src/views/Login.vue
@@ -1,6 +1,5 @@
 <template>
   <v-app>
-  <meta name="google-signin-client_id" content="761915308223-ua9pjnk2765b2qd88dda10htmhlu64js.apps.googleusercontent.com">
     <v-main class="grey lighten-2">
       <h1>Welcome to the {{'PLACEHOLDER-TEXT'}} Church Management System.</h1>
       <v-container>
@@ -40,8 +39,9 @@
           <v-col>
           </v-col>
           <v-col>
-            <button v-on:click="login()" class="btn btn-dark btn-lg " style="width:20%">Google</button>
-            <div class="g-signin2" data-onsuccess="onSignIn" data-onfailure="failed" data-theme="dark" style="margin: 2% auto; width:200px"></div>
+            <v-sheet class="grey lighten-2">
+              <div class="g-signin2" data-onsuccess="onSignIn" data-onfailure="failed" data-theme="dark" style="margin: 2% auto; width:100px;"></div>
+            </v-sheet>
           </v-col>
           <v-col>
           </v-col>
@@ -50,8 +50,6 @@
     </v-main>
   </v-app>
 </template>
-
-<script src="https://apis.google.com/js/platform.js" async defer></script>
 
 <script>
 export default {
@@ -65,10 +63,20 @@ export default {
       this.$router.push("/");
     }
   },
-}
-window.onLoadCallback = function(){
-  gapi.auth2.init({
-    client_id: '761915308223-ua9pjnk2765b2qd88dda10htmhlu64js.apps.googleusercontent.com'
-  });
+   mounted(){
+      let recaptchaScript = document.createElement('script')
+      recaptchaScript.setAttribute('src', 'https://apis.google.com/js/platform.js')
+      document.head.appendChild(recaptchaScript)
+      
+      this.polling = setInterval(() => {
+      if(window.user){
+        this.user = window.user;
+        this.$router.push("/");
+      }
+    }, 1000)
+  },
+  beforeDestroy() {
+    clearInterval(this.polling);
+  }
 }
 </script>
